+++ changelog.md
@@ -3,6 +3,7 @@
 ## Bugfixes
 
 - *Addon should now always properly initialize when reloading as the ML.*
+- *No longer prints the reroll message for candidates not being asked to reroll.*
 
 # 3.13.2